removed incorrect reference
The matrix passed to operator* should be
copied by value, not by reference, so that
the original matrix is not modified. For a real
working code, however, it should be a const
reference but that would require further
memory management which would obstruct
what is to be shown here (testing), thus, a
simple copy suffice here.